:truck: move event_handling folder to services
