+++ app/services/onboarding/util/register_issuer_did.py
@@ -2,10 +2,10 @@ from logging import Logger
 
 from aries_cloudcontroller import AcaPyClient, InvitationCreateRequest, InvitationRecord
 
-from app.event_handling.sse_listener import SseListener
 from app.exceptions import CloudApiException
 from app.services import acapy_ledger, acapy_wallet
 from app.services.acapy_wallet import Did
+from app.services.event_handling.sse_listener import SseListener
 from app.services.onboarding.util.set_endorser_metadata import (
     set_author_role,
     set_endorser_info,
