+++ app/tests/e2e/test_credentials.py
@@ -2,9 +2,9 @@ import time
 
 import pytest
 
-from app.event_handling.sse_listener import SseListener
 from app.models.tenants import CreateTenantResponse
 from app.routes.issuer import router
+from app.services.event_handling.sse_listener import SseListener
 from app.tests.util.ecosystem_connections import FaberAliceConnect
 from shared import RichAsyncClient
 from shared.models.credential_exchange import CredentialExchange
